Remove reset local votes when investing
When investor buys more shares, this resets the local votes
Wasn't causing visible issues as we use only voted round to prevent double voting
but inconsistent
this state may be removed later as it's likely not needed

Some tests failing as they use old functionality to retrieve app state directly and None/0 is different. Will be fixed in next commit.