Verify that withdrawal notes have prefix
We also switched to using a project uuid (client generated) instead of the backend id, since we need the id to verify the note prefix in teal (see note in teal for why it's needed) and at the moment when the teal is redendered, there's no backend id yet (it's retuned when saving the generated programs).

The uuid has other benefits, e.g. to store the project data on-chain. The idea is to remove entirely the backend and use only uuid to identify projects.