Implement global platform fee and capi token
Holding $capi entitles to collect dividend (a share of the global fee) from all the DAOs.
The mechanism is basically the same as in the DAOs (using an app to remember the global total collected and already collected local state, a dedicated escrow and token).
The global fee is charged as part of the DAO's draining transaction group. This means that in order to make funds retrievable by anyone in the DAO (to the creator or dividends to the investors), the global fee has to be paid.