Don't allow to set the setup global state 2x
No particular reason other than this data is intended to be part of the TEAL - it can't be for technical reasons, because when the app is created the escrows (which need the app id) aren't complete yet.
And the closest to having the addresses in code (using state) is state that can be set only once.