Remove "weird branching" comment from PyTeal
This flow seems correct for ~1:1 with the old TEAL

Alternative / possible improvement:
Add an identifying argument to all the groups, so we branch based on that and group size check becomes assert
The idea would be to check non-argument based conditions first (e.g. Txn.application_id() == Int(0)) and "the rest" is expected to always have 1 argument, this way no additional branching needed for arguments or group size.