Improve types for dao app id and dao id
Dao id is also a wrapper for dao app id - the app id can be used as dao identifier, since it's permanent and unique on the blockchain. We prefer passing a more general "dao id" to contexts that just want a dao id, thus wrap the app id in the dao id for those contexts.